Creating a Template Project
This is a template to every new game that will be added to Remar.
The basic changes (such as Spring Security configs, dependencies, Mongo, materialize, etc) are all done here,
So when creating a new project to add into Remar, we have a base project to copy from to start working on what you really need to and not waste your time with basic configurations.