Automatically creating the game DataBase when adding
Whenever a new game would be added, we needed to manually create a database. We wanted to automate this so we dont need to create those databases manually anymore.
Another thing that was done is putting the .env file in gitignore so we dont show the passwords, and now we maintain the sample file inside github

- In order to make the automatic 'adding database'  work, what was done:
  - remar container in docker now has the env vars for user and passwords for the databases (.env)
  - at the Resource Controller now when approving the game, we call the script "create_game_database" which will call the mysql to create the database